lab7_1.cpp completed

+++ lab7_1.cpp
@@ -2,9 +2,9 @@
 using namespace std;
 
 int main() {
-    int a  =  20; 
-    int b = 30;
-    int c = 4;
+    double a  =  20; 
+    double b = 30;
+    double c = 4;
     double output;
 
     output = a - b/c;
